bot_backend/requirements.txt
django==4.0.4
requests==2.27.1
django-environ==0.8.1
gunicorn==20.1.0
python-telegram-bot==13.11
ipython==8.3.0
emoji==1.7.0
Pillow==9.2.0